Pagination + junit tests

+++ src/main/java/com/excilys/formation/java/service/impl/ComputerDBServiceImpl.java
@@ -12,14 +12,14 @@ import com.excilys.formation.java.persistence.DaoFactory;
 import com.excilys.formation.java.service.ComputerDBService;
 
 public enum ComputerDBServiceImpl implements ComputerDBService {
-  
+
   INSTANCE;
 
-  private Logger logger = LoggerFactory.getLogger(ComputerDBServiceImpl.class);
-  
-  private ComputerDao                        computerDao;
+  private Logger      logger = LoggerFactory.getLogger(ComputerDBServiceImpl.class);
 
-  private DaoFactory                         daoFactory;
+  private ComputerDao computerDao;
+
+  private DaoFactory  daoFactory;
 
   /**
    * Singleton : provide the access service to the database
@@ -39,6 +39,10 @@ public enum ComputerDBServiceImpl implements ComputerDBService {
    */
   @Override
   public Computer getOne(Long id) {
+    if (id == null || id < 1) {
+      logger.error("Error with getOne()");
+      throw new IllegalArgumentException("id cannot be null or negative");
+    }
     return computerDao.getOne(id);
   }
 
@@ -48,7 +52,7 @@ public enum ComputerDBServiceImpl implements ComputerDBService {
   @Override
   public void create(Computer c) {
     if (c == null) {
-      logger.error("SQLError with create()");
+      logger.error("Error with create()");
       throw new IllegalArgumentException("Computer cannot be null");
     }
     computerDao.create(c);
@@ -60,7 +64,7 @@ public enum ComputerDBServiceImpl implements ComputerDBService {
   @Override
   public void update(Computer c) {
     if (c == null) {
-      logger.error("SQLError with update()");
+      logger.error("Error with update()");
       throw new IllegalArgumentException("Computer cannot be null");
     }
     computerDao.update(c);
@@ -71,9 +75,9 @@ public enum ComputerDBServiceImpl implements ComputerDBService {
    */
   @Override
   public void delete(Long id) {
-    if (id == null) {
-      logger.error("SQLError with delete()");
-      throw new IllegalArgumentException("Id cannot be null");
+    if (id == null || id < 1) {
+      logger.error("Error with delete()");
+      throw new IllegalArgumentException("Id cannot be null or negative");
     }
     computerDao.delete(id);
   }
@@ -83,6 +87,10 @@ public enum ComputerDBServiceImpl implements ComputerDBService {
    */
   @Override
   public Page<Computer> createPage(Page<Computer> page) {
+    if (page == null) {
+      logger.error("Error with createPage()");
+      throw new IllegalArgumentException("Page cannot be null");
+    }
     return computerDao.createPage(page);
   }
